requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
praw==7.7.1
python-dotenv==1.0.1
openai>=1.0.0
pydantic==2.6.0
